Revert "Merge develop"
This reverts commit 40798fbe72cf037e6c300a560161c2a3fbc2b5ad, reversing
changes made to 5e2d9048796fb0ba4379f73efa9251cdf60ae298.
