+++ app/src/main/AndroidManifest.xml
@@ -5,8 +5,8 @@
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <application
-        android:name=".CloudApplication"
         android:usesCleartextTraffic="true"
+        android:name=".CloudApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
