--- app/src/main/java/com/knu/cloud/model/instanceCreate/Flavor.kt
@@ -1,11 +1,0 @@
-package com.knu.cloud.model.instanceCreate
-
-data class Flavor (
-    val name: String,
-    val vcpus: Int,
-    val ram: Int,
-    val diskTotal: Int,
-    val rootDisk: Int,
-    val ephemeralDisk: Int,
-    val public: String
-)
