--- app/src/main/java/com/knu/cloud/model/instanceCreate/Keypair.kt
@@ -1,6 +1,0 @@
-package com.knu.cloud.model.instanceCreate
-
-data class Keypair(
-    val name: String,
-    val type: String,
-)
