--- app/src/main/java/com/knu/cloud/model/instanceCreate/Network.kt
@@ -1,9 +1,0 @@
-package com.knu.cloud.model.instanceCreate
-
-data class Network(
-    val network: String,
-    val subNet: String,
-    val share: String,
-    val adminState: String,
-    val state: String
-)
