--- app/src/main/java/com/knu/cloud/model/instanceCreate/Source.kt
@@ -1,9 +1,0 @@
-package com.knu.cloud.model.instanceCreate
-
-data class Source(
-    val name: String,
-    val update: String,
-    val size: Double,
-    val format: String,
-    val visible: String
-)
